# Documents/SUNNU_MBAYE_MI/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/listeCategorie": {
            "get": {
                "summary": "Liste de toutes les catégories",
                "description": "Liste toutes les catégories.",
                "operationId": "1b5bbfda02b549f42ad1aff814f9c7f9",
                "responses": {
                    "200": {
                        "description": "Liste des catégories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categorie": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Categorie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/api/voiplusCategorie/{id}": {
            "get": {
                "summary": "Détails d'une catégorie",
                "description": "Afficher les détails d'une catégorie.",
                "operationId": "d2f2a07544a4cb4ccaa560f13fbd63bc",
                "parameters": [
                    {
                        "name": "categori_id",
                        "in": "path",
                        "description": "ID de la catégorie à afficher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categorie": {
                                            "$ref": "#/components/schemas/Categorie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/AjoutCategorie": {
            "post": {
                "summary": "Ajouter une catégorie",
                "description": "Ajouter une nouvelle catégorie.",
                "operationId": "8e6f3dac358c84ecbc6d9b71179bd598",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_categories": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie ajoutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie ajoutée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/modifieCategorie/{id}": {
            "put": {
                "summary": "Modifier une catégorie",
                "description": "Modifier une catégorie existante.",
                "operationId": "865ac9f0c75b36aa7f51300674f6513e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_categories": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie modifiée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie modifiée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supCategorie/{id}": {
            "delete": {
                "summary": "Supprimer une catégorie",
                "description": "Supprimer une catégorie.",
                "operationId": "a1e7e68092878b5fd5d15063c05f9de2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}